Fix hot reload
