+++ app/TabContext.tsx
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+
+export const TabContext = createContext<
+  | {
+      setTab: (tab: string) => void;
+    }
+  | undefined
+>(undefined);
