+++ app/stringtoKiloBytes.ts
@@ -1,0 +1,3 @@
+export function stringToKiloBytes(data: string) {
+  return ((encodeURI(data).split(/%..|./).length - 1) / 1024).toFixed(2);
+}
